pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-generativeai>=0.8.5",
    "python-dotenv>=1.2.1",
    "streamlit-nightly>=1.51.1.dev20251124",
]
